Moved gds_wait_request related functions to mlx5-exp
