+++ src/apis.cpp
@@ -56,7 +56,7 @@
 
 //-----------------------------------------------------------------------------
 
-static void gds_init_ops(struct peer_op_wr *op, int count)
+void gds_init_ops(struct peer_op_wr *op, int count)
 {
         int i = count;
         while (--i)
@@ -80,13 +80,15 @@ static void gds_init_send_info(gds_send_request_t *info)
 
 static void gds_init_wait_request(gds_wait_request_t *request, uint32_t offset)
 {
+        gds_mlx5_exp_wait_request_t *gmexp_request;
         gds_dbg("wait_request=%p offset=%08x\n", request, offset);
         memset(request, 0, sizeof(*request));
-        request->peek.storage = request->wr;
-        request->peek.entries = sizeof(request->wr)/sizeof(request->wr[0]);
-        request->peek.whence = IBV_EXP_PEER_PEEK_ABSOLUTE;
-        request->peek.offset = offset;
-        gds_init_ops(request->peek.storage, request->peek.entries);
+
+        request->dtype = GDS_DRIVER_TYPE_MLX5_EXP;
+
+        gmexp_request = to_gds_mexp_wait_request(request);
+
+        gds_mlx5_exp_init_wait_request(gmexp_request, offset);
 }
 
 //-----------------------------------------------------------------------------
@@ -273,7 +275,9 @@ int gds_stream_post_send_all(CUstream stream, int count, gds_send_request_t *req
 
 int gds_prepare_wait_cq(struct gds_cq *cq, gds_wait_request_t *request, int flags)
 {
-        int retcode = 0;
+        gds_mlx5_exp_cq_t *gmexpcq;
+        gds_mlx5_exp_wait_request_t *gmexp_request;
+
         if (flags != 0) {
                 gds_err("invalid flags != 0\n");
                 return EINVAL;
@@ -281,51 +285,19 @@ int gds_prepare_wait_cq(struct gds_cq *cq, gds_wait_request_t *request, int flag
 
         gds_init_wait_request(request, cq->curr_offset++);
 
-        retcode = ibv_exp_peer_peek_cq(cq->cq, &request->peek);
-        if (retcode == -ENOSPC) {
-                // TODO: handle too few entries
-                gds_err("not enough ops in peer_peek_cq\n");
-                goto out;
-        } else if (retcode) {
-                gds_err("error %d in peer_peek_cq\n", retcode);
-                goto out;
-        }
-        //gds_dump_wait_request(request, 1);
-        out:
-	       return retcode;
+        gmexpcq = to_gds_mexp_cq(cq);
+        gmexp_request = to_gds_mexp_wait_request(request);
+
+        return gds_mlx5_exp_prepare_wait_cq(gmexpcq, gmexp_request, flags);
 }
 
 //-----------------------------------------------------------------------------
 
 int gds_append_wait_cq(gds_wait_request_t *request, uint32_t *dw, uint32_t val)
 {
-        int ret = 0;
-        unsigned MAX_NUM_ENTRIES = sizeof(request->wr)/sizeof(request->wr[0]);
-        unsigned n = request->peek.entries;
-        struct peer_op_wr *wr = request->peek.storage;
+        gds_mlx5_exp_wait_request_t *gmexp_request = to_gds_mexp_wait_request(request);
 
-        if (n + 1 > MAX_NUM_ENTRIES) {
-            gds_err("no space left to stuff a poke\n");
-            ret = ENOMEM;
-            goto out;
-        }
-
-        // at least 1 op
-        assert(n);
-        assert(wr);
-
-        for (; n; --n) wr = wr->next;
-        assert(wr);
-
-        wr->type = IBV_EXP_PEER_OP_STORE_DWORD;
-        wr->wr.dword_va.data = val;
-        wr->wr.dword_va.target_id = 0; // direct mapping, offset IS the address
-        wr->wr.dword_va.offset = (ptrdiff_t)(dw-(uint32_t*)0);
-
-        ++request->peek.entries;
-
-        out:
-        return ret;
+        return gds_mlx5_exp_append_wait_cq(gmexp_request, dw, val);
 }
 
 //-----------------------------------------------------------------------------
@@ -346,12 +318,16 @@ int gds_stream_post_wait_cq_all(CUstream stream, int count, gds_wait_request_t *
 
 static int gds_abort_wait_cq(struct gds_cq *cq, gds_wait_request_t *request)
 {
+        gds_mlx5_exp_cq_t *gmexpcq;
+        gds_mlx5_exp_wait_request_t *gmexp_request;
+
         assert(cq);
         assert(request);
-        struct ibv_exp_peer_abort_peek abort_ctx;
-        abort_ctx.peek_id = request->peek.peek_id;
-        abort_ctx.comp_mask = 0;
-        return ibv_exp_peer_abort_peek_cq(cq->cq, &abort_ctx);
+
+        gmexpcq = to_gds_mexp_cq(cq);
+        gmexp_request = to_gds_mexp_wait_request(request);
+
+        return gds_mlx5_exp_abort_wait_cq(gmexpcq, gmexp_request);
 }
 
 //-----------------------------------------------------------------------------
@@ -549,7 +525,7 @@ static int calc_n_mem_ops(size_t n_descs, gds_descriptor_t *descs, size_t &n_mem
                         n_mem_ops += desc->send->commit.entries + 2; // extra space, ugly
                         break;
                 case GDS_TAG_WAIT:
-                        n_mem_ops += desc->wait->peek.entries + 2; // ditto
+                        n_mem_ops += gds_mlx5_exp_get_num_wait_request_entries(to_gds_mexp_wait_request(desc->wait)) + 2; // ditto
                         break;
                 case GDS_TAG_WAIT_VALUE32:
                 case GDS_TAG_WRITE_VALUE32:
@@ -618,15 +594,15 @@ int gds_stream_post_descriptors(CUstream stream, size_t n_descs, gds_descriptor_
                         break;
                 }
                 case GDS_TAG_WAIT: {
-                        gds_wait_request_t *wreq = desc->wait;
+                        gds_mlx5_exp_wait_request_t *wreq = to_gds_mexp_wait_request(desc->wait);
                         int flags = 0;
                         if (move_flush && i != last_wait) {
                                 gds_dbg("discarding FLUSH!\n");
                                 flags = GDS_POST_OPS_DISCARD_WAIT_FLUSH;
                         }
-                        retcode = gds_post_ops(peer, wreq->peek.entries, wreq->peek.storage, params, flags);
+                        retcode = gds_mlx5_exp_stream_post_wait_descriptor(peer, wreq, params, flags);
                         if (retcode) {
-                                gds_err("error %d in gds_post_ops\n", retcode);
+                                gds_err("error %d in gds_mlx5_exp_stream_post_wait_descriptor\n", retcode);
                                 ret = retcode;
                                 goto out;
                         }
@@ -697,10 +673,10 @@ int gds_post_descriptors(size_t n_descs, gds_descriptor_t *descs, int flags)
                 }
                 case GDS_TAG_WAIT: {
                         gds_dbg("desc[%zu] WAIT\n", i);
-                        gds_wait_request_t *wreq = desc->wait;
-                        retcode = gds_post_ops_on_cpu(wreq->peek.entries, wreq->peek.storage, flags);
+                        gds_mlx5_exp_wait_request_t *wreq = to_gds_mexp_wait_request(desc->wait);
+                        retcode = gds_mlx5_exp_post_wait_descriptor(wreq, flags);
                         if (retcode) {
-                                gds_err("error %d in gds_post_ops_on_cpu\n", retcode);
+                                gds_err("error %d in gds_mlx5_exp_post_wait_descriptor\n", retcode);
                                 ret = retcode;
                                 goto out;
                         }
